added presence command

+++ bot.js
@@ -1,4 +1,5 @@
 const Discord = require('discord.js');
+const { TIMEOUT } = require('dns');
 const fetch = require('node-fetch');
 const client = new Discord.Client();
 const { token, ownerid } = require('./config.json');
@@ -13,10 +14,18 @@ client.once('ready', () => {
 client.on('message', message => {
   const args = message.content.slice(prefix.length).trim().split(/ +/);
   const command = args.shift().toLowerCase();
-  console.log(`From: ${message.author.tag} > ${message.content}`)
+   if(message.guild === null) {
+	  console.log(`DM From: ${message.author.tag} > ${message.content}`)
+	  if(message.content.startsWith(prefix)) {
+		  message.author.send('Commands can only be run from a server, not a dm.')
+	  }
+	  return
+  } else {
+	  console.log(`From: ${message.author.tag} > ${message.content}`)
+  }
   if(message.author.bot) return
-  if(message.guild === null) return
-if (message.content === '/ping') {
+ 
+  if (message.content === '/ping') {
 	message.channel.send('Pong.');
 } else if (message.content === `${prefix}count`) {
 	message.channel.send(`Current member count: ${message.guild.memberCount}`);
@@ -34,9 +43,8 @@ if (message.content === '/ping') {
 	.setColor('#21B8FF')
 	.setTitle('Help Page')
 	.addFields(
-		{ name: 'Commands:', value: '**help** - displays this embed\n**echo** - echos what you write\n**delete** - deletes your message\n**serverinfo** - displays connection info along with the status and dynmap\n**count** - displays member count\n**userinfo** - displays username and id' },
-		{ name: 'Abilities:', value: 'Hating Nubia'},
-		{ name: 'Debugging:', value: '**args** - sends the arguments in your message\n**argsinfo** - sends command and arguments\n**shutdown** - shuts down the bot'}
+		{ name: 'Commands:', value: '**help** - displays this embed\n**echo** - echos what you write\n**delete** - deletes your message\n**serverinfo** - displays connection info along with the status and dynmap\n**count** - displays member count\n**userinfo** - displays username and id\n**slowmode** - sets slowmode to specified seconds' },
+		{ name: 'Abilities:', value: 'Hating Nubia'}
 	)
   message.channel.send(helpembed)
 } else if (command === "debug") {
@@ -48,8 +56,8 @@ if (message.content === '/ping') {
 	)
 	message.channel.send(debugembed);
 } else if (command === 'spam') {
-if (!args.length) {
-  return message.channel.send('u didnt say what to spam')
+  if (!args.length) {
+  		return message.channel.send('u didnt say what to spam')
   }
   const spam = args[0]
   message.delete()
@@ -61,6 +69,9 @@ if (!args.length) {
 } else if (command === 'delete') {
 	message.delete()
 } else if (command === 'echo') {
+	if (!args.length) {
+		return message.channel.send('you didnt say anything to echo')
+	}
 	message.delete()
 	message.channel.send(message.content.replace("/echo", ""));
 } else if (command === 'serverinfo') {
@@ -103,6 +114,63 @@ if (!args.length) {
 	.setTitle('Github')
 	.setDescription('https://github.com/johng3587/KineticSMPBot')
 	message.channel.send (githubembed)
+} else if (command === 'slowmode') {
+	if (!args.length) {
+		 return message.channel.send('you didnt say a slowmode time')
+	} else if (!message.member.hasPermission('MANAGE_CHANNELS')) {
+		message.channel.send ('you dont have permission to do that')
+	} else {
+		message.channel.setRateLimitPerUser(args[0], `${message.author.tag} requested a slowmode of ${args[0]} seconds.`)
+		message.channel.send(`Success! Slowmode set to ${args[0]} seconds.`)
+	}
+} else if (command === 'log') {
+	if (!args.length) {
+		return message.channel.send('You didnt say anything to log!')
+	} else if (message.author.id !== ownerid) {
+		message.channel.send ('Only the bot owner can log messages to the console.')
+	} else {
+		const logcontent = message.content
+		console.log('\x1b[32m\x1b[4m\x1b[1m%s', `Custom log: ${message.content.replace("/log ", "")}`);
+		message.channel.send('Message logged!')
+	}
+} else if (command === 'game') {
+	if (!args.length) {
+		message.channel.send ('you need to name a status')
+	} else if (message.author.id !== ownerid) {
+		message.channel.send ('Only the bot owner can change my status.')
+	} else {
+		client.user.setActivity(message.content.replace("/game ", ""))
+		message.channel.send(`Status set to ${message.content.replace("/game ", "")}`)
+	}
+	 
+} else if (command === 'setavatar') {
+	if (!args.length) {
+		message.channel.send ('you need to rovide a url')
+	} else if (message.author.id !== ownerid) {
+		message.channel.send ('Only the owner can change my profile picture')
+	} else {
+		client.user.setAvatar(args[0])
+	}
+} else if (command === 'presence') {
+	if (!args.length) {
+		message.channel.send ('You need to provide a presence. Allowed values are:\nonline, idle, invisible, dnd')
+	} else if (message.author.id !== ownerid) {
+		message.channel.send('Only the owner can change my presence')
+	} else if (args[0] === 'online') {
+		client.user.setPresence({ status: 'online' })
+		message.channel.send(`Presence set to ${args[0]}`)
+	} else if (args[0] === 'idle') {
+		client.user.setPresence({ status: 'idle' })
+		message.channel.send(`Presence set to ${args[0]}`)
+	} else if (args[0] === 'invisible') {
+		client.user.setPresence({ status: 'invisible' })
+		message.channel.send(`Presence set to ${args[0]}`)
+	} else if (args[0] === 'dnd') {
+		client.user.setPresence({ status: 'dnd' })
+		message.channel.send(`Presence set to ${args[0]}`)
+	} else {
+		message.channel.send(`Invalid argument: ${args[0]}. Valid arguments are:\nonline, idle, invisible, dnd`)
+	}
 }
 
 });
